DFV Refactor: add a `_src/fields/...` directory
