+++ tests/js/pods-dfv/file-upload/models/file-upload-model.js
@@ -1,5 +1,5 @@
 /*global jQuery, _, Backbone, Mn, assert */
-import {FileUploadCollection, FileUploadModel} from 'pods-dfv/_src/file-upload/file-upload-model';
+import {FileUploadCollection, FileUploadModel} from 'pods-dfv/_src/fields/file-upload/file-upload-model';
 
 const default_model = new FileUploadModel();
