+++ tests/js/pods-dfv/pick/views/multiselect-selection-limits.js
@@ -1,9 +1,9 @@
 /*global assert */
 const inspect = require( 'util' ).inspect;
 
-import {SelectView} from 'pods-dfv/_src/pick/views/select-view';
+import {SelectView} from 'pods-dfv/_src/fields/pick/views/select-view';
 import {PodsDFVFieldModel} from 'pods-dfv/_src/core/pods-field-model';
-import {RelationshipCollection} from 'pods-dfv/_src/pick/relationship-model';
+import {RelationshipCollection} from 'pods-dfv/_src/fields/pick/relationship-model';
 
 const collection = new RelationshipCollection( [
 	{ id: 0, name: 'zero' },
