+++ tests/js/pods-dfv/pick/views/select-view.js
@@ -1,7 +1,7 @@
 /*global assert */
-import {SelectView} from 'pods-dfv/_src/pick/views/select-view';
+import {SelectView} from 'pods-dfv/_src/fields/pick/views/select-view';
 import {PodsDFVFieldModel} from 'pods-dfv/_src/core/pods-field-model';
-import {RelationshipCollection} from 'pods-dfv/_src/pick/relationship-model';
+import {RelationshipCollection} from 'pods-dfv/_src/fields/pick/relationship-model';
 
 let pickFormatTypes = [
 	{ type: 'single', isMultiple: false },
