+++ ui/js/pods-dfv/_src/field-manifest.js
@@ -1,3 +1,3 @@
-export { FileUpload } from 'pods-dfv/_src/file-upload/file-upload';
-export { Pick } from 'pods-dfv/_src/pick/pick';
-export { PodsDFVText } from 'pods-dfv/_src/text/text';
+export { FileUpload } from 'pods-dfv/_src/fields/file-upload/file-upload';
+export { Pick } from 'pods-dfv/_src/fields/pick/pick';
+export { PodsDFVText } from 'pods-dfv/_src/fields/text/text';
