+++ ui/js/pods-dfv/_src/fields/file-upload/uploaders/media-modal.js
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
 /*global jQuery, _, Backbone, Marionette, wp, PodsI18n */
-import { PodsFileUploader } from 'pods-dfv/_src/file-upload/uploaders/pods-file-uploader';
+import { PodsFileUploader } from 'pods-dfv/_src/fields/file-upload/uploaders/pods-file-uploader';
 
 export const MediaModal = PodsFileUploader.extend( {
 	mediaObject: {},
