+++ ui/js/pods-dfv/_src/fields/file-upload/views/file-upload-form.js
@@ -1,5 +1,5 @@
 /*global jQuery, _, Backbone, Marionette, wp */
-import template from 'pods-dfv/_src/file-upload/views/file-upload-form.html';
+import template from 'pods-dfv/_src/fields/file-upload/views/file-upload-form.html';
 
 import { PodsFieldView } from 'pods-dfv/_src/core/pods-field-views';
