+++ ui/js/pods-dfv/_src/fields/file-upload/views/file-upload-item.html
