+++ ui/js/pods-dfv/_src/fields/file-upload/views/file-upload-list.js
@@ -1,5 +1,5 @@
 /*global jQuery, _, Backbone, Marionette, wp */
-import template from 'pods-dfv/_src/file-upload/views/file-upload-item.html';
+import template from 'pods-dfv/_src/fields/file-upload/views/file-upload-item.html';
 
 import { PodsFieldListView, PodsFieldView } from 'pods-dfv/_src/core/pods-field-views';
