+++ ui/js/pods-dfv/_src/fields/file-upload/views/file-upload-queue.js
@@ -1,5 +1,5 @@
 /*global jQuery, _, Backbone, Marionette, wp */
-import template from 'pods-dfv/_src/file-upload/views/file-upload-queue.html';
+import template from 'pods-dfv/_src/fields/file-upload/views/file-upload-queue.html';
 
 export const FileUploadQueueModel = Backbone.Model.extend( {
 	defaults: {
