+++ ui/js/pods-dfv/_src/fields/pick/views/add-new.html
@@ -1,4 +1,4 @@
-<a href="/wp-admin/?pods_modal=1"
+<a href="/wp-admin?pods_modal=1"
 		class="button pods-related-add-new pods-modal"
 		data-pod-id="<%- fieldConfig.pod_id %>"
 		data-field-id="<%- fieldConfig.id %>"
