+++ ui/js/pods-dfv/_src/fields/pick/views/checkbox-item.html
