+++ ui/js/pods-dfv/_src/model-manifest.js
@@ -1,3 +1,3 @@
 export { PodsDFVFieldModel } from 'pods-dfv/_src/core/pods-field-model';
-export * from 'pods-dfv/_src/pick/relationship-model';
-export * from 'pods-dfv/_src/file-upload/file-upload-model';
+export * from 'pods-dfv/_src/fields/pick/relationship-model';
+export * from 'pods-dfv/_src/fields/file-upload/file-upload-model';
